Multi-Language support. (#17)
* WIP. Add better Text Logic.

The progress bar time display does not work yet. That's why it is in a different branch as well.

* Fix finally the text positions. :stuck_out_tongue:

* Why do I need this again..?

It's annoying, because otherwise it would crash...

* Add better Batterys. Thanks to TotallyNotGuy! :stuck_out_tongue:

* Show the progressbar in the Music Player again.

* A Few Optimizations for the Music Player.

* Add a Second Settings Page!

You can change the Bubble color and enable/disable the Animation now!

* Add Battery Percent Shown/Hidden as an Option.

* Add Geometry animated Texture!

* small fixes + Optimizations. :stuck_out_tongue:

* Oopsie.. :stuck_out_tongue:

* Comment out external Image stuff for now, because this is not implemented yet.

So people don't try, to get a Crash because of that. :stuck_out_tongue:

* Cleanup of the spritesheet stuff.

Now also generate the animation spritesheet at start!

* animation spritesheet cleanup!

* Add a second Overlay.

You can switch to it in the Settings.

* Oopsie..

* Add a Back button.

* Fix text positions.

* Fix Messages disappearing after 2-3 seconds.

* Add pkmn-chest to the download screen.

And hopefully this will fix the Screen Issue, what happens sometimes.

* WIP. Thanks to FlagBrew and PKSM!

Multi Language support! :stuck_out_tongue:

* PKMN-Chest Release can be downloaded now!

The Progressbar is broken on it, so I changed it.

* Main Menu - Done.

And add Translation Strings for every supported Language!

* Next Screen -> Done.

* Next Strings!

* Begin to Translate Music Player Strings. **Not Done Yet!**

* Add Settings Screen. -> Next work tomorrow/later.

Also the strings doesn't work yet for the Settings Screen `std::string`.

* Never Translate `std::string`! This cause to a `missing` String.

* Add Updater Screen Strings.

* Next Strings! Now : Add the other Missing Strings.