Most stuff from the Graphic-Rewrite. (#31)
* Begin with the 3DS Graphic-Rewrite!

After this is done and merged, the Graphics looks better than before!

* Merge branch 'master' into graphic-rewrite

* Remove unneded Screenshot Feature.

Because Luma3DS can do Screenshots as well!

* Add the default 3DS Color to the DSi Version as well!

Also use enums instead of definitions for the Screen Modes.

* [3DS] : Some Text Centering. :stuck_out_tongue:

This only center the Text on the Top Bar or Bottom Bar on the Top Screen.

* [DSi] : Attempt to try to add a Config File.

It doesn't save the ini file yet, but it Creates the Folder for it.

* [DSi] : The Config Stuff works now.

* [DSi] : Add WIP Settings Screen.

You can't really do anything on it yet. :stuck_out_tongue:

* [DSi] : Add touch to Main Menu and begin to add Text.

Also prepare the Keyboard Graphic, because I likely make a usage of it in the next Commits.

* Add those Changes from master to the current Branch as well.

* DSi: Move input code into screen modes

ALSO: Add bool for if screen is drawn

* Fix compile conflict with enum and screen.

* Fix Screen Flicker by touching the Screen.

Now it only redraw, if the Touch position with the button is valid. ;P

* It probably may be better, to add this, so it redraw the screen after the color has changed.

* [3DS] : Add new Buttons and change Icons for the new Buttons.

* Fix fileBrowse Crash.

* Next Graphic-Rewrite work! :stuck_out_tongue:

It's still a work in progress and it takes it's time.

* Get rid of the Roboto Font (FileBrowse) And use the System one.

piepie62 suggested that the Roboto Font would be bad because it miss Chars, so I will use the System Font for it. Thank you!

* Add Screen Classes to the DSi Version.

* [Switch] : Add Screen Classes as well.

* [DSi] : Redo the Classes.

I hope this is correct?

* [DSi] : Finish Screen Classes for now.

* [Switch] : Do the same work with the classes here as well.

* [3DS] : Prepare the Screen Class work for the next work.

I'll take a break for today and work probably tomorrow on it again.

* Add VSC build task and fix VSC intellisense errors

* gitignore the t3x files as they change when built

* Screen-Classes. (#30)

* remove azure pipeline for now, until this is finished.

* Do a new start with Universal-Manager 3DS.

First Commit : Add the Main Menu.

* MainMenu Class -> Done. (1/30)

* Add the whole Contents back.

Also : FileManager Class work.

* Step 1 : Add all the classes -> Done.

Next Step : Fix some stuff.

* Fix FTP Exit freeze.

* Image Viewer fix. (Currently workaround.)

* Merge the Updater Screens into one class.

* Exit in the Main Menu with Start.

* change `button.hpp` to `structs.hpp`.

* Merge Script Stuff to one class.

* Some Class Stuff.

* Merge Text Editor to one class.

* Some changes in the settings.

* Merge all the Music Stuff into one class!

To-Do : Fix Playlist not playing.

* [3DS & Switch] : Some work.

* [3DS] : Fully implement back the helper Boxes.

* Add LeafEdit Nightly to the Updater Screen.

* Add azure again. :stuck_out_tongue:

* Show git commit sha / tag instead of hardcoded version.

* Add Placeholder for LeafEdit Release.

* Fix restarting after selfupdate?

It didn't restarted after self updating.

* Fix Updater Screen CAPS on some Buttons.

And also reposition the Text a little bit, so it isn't directly at the begin of the Button.

* Add Nightly SHA to Switch & DS(i) Version as well.

Also update the Github Templates.

* [3DS] : Install all CIA's from TWiLightMenu Nightly again!

I forgot it. :stuck_out_tongue:

* Add LeafEdit Release Download Function!

* Add Animated Selector for UM-3DS FileBrowse as well!

The Selector has the Bar Color Btw.

* Now the Text of the selected File can be seen on every BG Color!

Only the Selected File though. :stuck_out_tongue:

* Add Fade Effect for GUI Button Selector.

Fade in the Bar Color as well! [Only Scripts and Main Menu for now.]

* Fix Scripts Screen Slowdowns.

* 'some' 3DS work.

* I did an Oopsie.

* Fix FileBrowse Slowdowns.

* Make the Black Selector with White Text as the Default one.

* This should fix the Issues with the Music Player?

* Fix Tic-Tac-Toe.

* Fix Tic-Tac-Toe really.

* Merge FileManager Options to FileManager Screen Class.

This should also fix the Crashes on the FileManager now.

* This should fix CIA Installing with the FileManager.

**USE AT YOUR OWN RISK!!!**

* Fix my Mistakes in the Updating TWLMenu Nightly part.

* Implement Fast Mode for the FileBrowse.

You can enable Fastmode with `R`! You can also disable it with `L`.

* Reduce the FileBrowse Selector Animation speed.

It was really too fast somehow. :stuck_out_tongue:

* FileManager work. :stuck_out_tongue:

* WIP Numpad.

* Get rid of old Text Function and use the new one.

* Fix `\n` by the Scripts.

* Starting to implement Multi Language support.

* Revert latest Commit [Multi Language Support].

This reverts commit 3de12b67803a5cf470e9d9000ab7a6e338756fb5.

* Music Sub Menu Graphic work.

* Updater Graphic work.

* Get rid of old buttons.

* Slow down the speed of the FileManager Options Selector.

It was too fast still.

* Add latest Text changes from LeafEdit here as well.

* Use `1` instead of `4` Sprite Drawing Functions.

* Add empty File Creation to the File Manager.

* small Keyboard fixes on the Text Editor.

The Keyboard didn't really respond sometimes there.. so it was likely a mistake with the Touch position.

* Filemanager Touch support -> Done!

* Implement Background Music Playing back.

This is neat for the Text Editor for example.

* Fix missing Scroll SFX in filebrowses.

* Change Button Tester Logic.

* Change the Logic again Lmao.

Now it only displays the Button, when it is pressed.

* Set a specific Speed in Pong.

Also: User can choose the speed by yourself as well! Default -> 5. Possible: 1-9.

* WIP: Implement Calculator Layout.

Next work -> The Calculator Functions.

* Fix Scripts Creator. (Now it works properly.)

* DS(i) version Logic changes.

* Last work for the next days.

This will implement the new Pong Mode, where I was talking about in the Universal-Server. How you can active this mode? Find it out by yourself. ( ͡° ͜ʖ ͡°)

* Derp fix.